Configure broker without modifying any files
The broker application now accepts a SETTINGS_PATH environment variable
that points to an application configuration file. That file then points
to a database configuration file and a log file, as well as defining
other settings, e.g., the auth token.

This ensures that our job never modifies anything in the package: an
important trait, as it allows multiple jobs on the same machine to share
the package.

We also added sensible MySQL server defaults for the admin username and
port and updated the manifest template to use them.