Deploy without a CF
This is a great bosh MySQL release that might not always be deployed
alongisde a CloudFoundry. All CF components are still in the release,
but we want to create a manifest that disables them without breaking the
deployment.

The mysql-stub-spiff-3-node.yml can be used on a clean bosh-lite,
without any CF and it will just work. We'll try this on AWS next and
come back with manifest additions if required.

Also relates to cloudfoundry-incubator/cf-mysql-acceptance-tests#2